chore: fixed the flow and added amountToPay readonly var in PaymentComponent protocol to check the order remaining amount and if nil use the Payment.amount, and other small fixes
