+++ AdyenComponents/Issuer List/IssuerListComponent.swift
@@ -66,7 +66,7 @@ public final class IssuerListComponent: PaymentComponent, PresentableComponent, 
                 
                 let details = IssuerListDetails(paymentMethod: self.issuerListPaymentMethod,
                                                 issuer: issuer.identifier)
-                self.submit(data: PaymentComponentData(paymentMethodDetails: details, amount: self.payment?.amount, order: self.order))
+                self.submit(data: PaymentComponentData(paymentMethodDetails: details, amount: self.amountToPay, order: self.order))
                 listViewController.startLoading(for: listItem)
             }
